use stdint types

+++ src/largedoublelinkedlist.cpp
@@ -1,31 +1,31 @@
 #include <sealib/largedoublelinkedlist.h>
 #include <iostream>
 
-Sealib::LargeDoubleLinkedList::LargeDoubleLinkedList(unsigned int size) :
+Sealib::LargeDoubleLinkedList::LargeDoubleLinkedList(uint32_t size) :
     links(size * 2, 1),
-    current(size == 0 ? (unsigned int) -1 : 0) {
+    current(size == 0 ? (uint32_t) -1 : 0) {
 }
 
-unsigned int Sealib::LargeDoubleLinkedList::get() {
-    if (current != (unsigned int) -1) {
-        unsigned int retVal = current;
+uint32_t Sealib::LargeDoubleLinkedList::get() {
+    if (current != (uint32_t) -1) {
+        uint32_t retVal = current;
         remove(current);
         return retVal;
     } else {
-        return (unsigned int) -1;
+        return (uint32_t) -1;
     }
 }
 
-unsigned int Sealib::LargeDoubleLinkedList::remove(unsigned int idx) {
-    if (current == (unsigned int) -1) {  // empty list
+uint32_t Sealib::LargeDoubleLinkedList::remove(uint32_t idx) {
+    if (current == (uint32_t) -1) {  // empty list
         return current;
     }
 
-    long actualIdx = idx * 2;
+    int32_t actualIdx = idx * 2;
 
-    long prevIdx = actualIdx - links[actualIdx] * 2;
-    long n = links.size();
-    long mod = prevIdx % n;
+    int32_t prevIdx = actualIdx - links[actualIdx] * 2;
+    int32_t n = links.size();
+    int32_t mod = prevIdx % n;
     if (prevIdx < 0) {
         prevIdx = n + mod;
     } else {
@@ -35,19 +35,19 @@ unsigned int Sealib::LargeDoubleLinkedList::remove(unsigned int idx) {
         prevIdx = 0;
     }
     if (prevIdx == actualIdx) {  // last element
-        current = (unsigned int) -1;
+        current = (uint32_t) -1;
         return idx;
     }
     links[prevIdx + 1] += links[actualIdx + 1];
 
-    long nextIdx = actualIdx + links[actualIdx + 1] * 2;
+    int32_t nextIdx = actualIdx + links[actualIdx + 1] * 2;
     nextIdx = nextIdx % n;
     links[nextIdx] += links[actualIdx];
 
-    current = static_cast<unsigned int>(nextIdx / 2);
+    current = static_cast<uint32_t>(nextIdx / 2);
     return current;
 }
 
 bool Sealib::LargeDoubleLinkedList::isEmpty() {
-    return current == (unsigned int) -1;
+    return current == (uint32_t) -1;
 }
